fix rendering of the trees

+++ SimWorld/Utils/NSArray+Extensions.h
@@ -10,7 +10,7 @@
 
 @interface NSMutableArray (Extensions)
 
-- (void)fillWith:(id)obj andTimes:(int)amount;
+- (void)fillWith:(id)obj andTimes:(NSInteger)amount;
 - (void)fillWithFloat:(float)obj andTimes:(int)amount;
 
 @end
